Comprobaciones en SAcliente y habitacion
No se cuelen transfer nulos o con campos nulos

+++ src/integracion/habitaciones/dao/imp/DAOHabitacionImp.java
@@ -172,7 +172,7 @@ public class DAOHabitacionImp implements DAOHabitacion {
 				}
 
 				habitacion.setNumHabitacion(rowsHabitaciones.getInt("numhabitacion"));
-				habitacion.setPrecio(rowsHabitaciones.getInt("precio"));
+				habitacion.setPrecio(rowsHabitaciones.getDouble("precio"));
 		
 				listaHabitaciones.add(habitacion);
 			}
